chore(ui): Move Layout to deprecated subfolder
