--- packages/ui/src/components/Layout/Layout.css
@@ -1,3 +1,0 @@
-:where(.cui-layout) {
-	display: contents;
-}
