+++ packages/ui/src/deprecated/Layout/ThemeSchemeContext.ts
@@ -1,9 +1,12 @@
 import { createContext, useContext } from 'react'
 import { ThemeScheme } from './Types'
 
+/** @deprecated (no alternative) */
 export const ThemeSchemeContext = createContext<ThemeScheme>({})
+/** @deprecated (no alternative) */
 export const TitleThemeSchemeContext = createContext<ThemeScheme>({})
 
+/** @deprecated (no alternative) */
 export const useThemeScheme = ({
    scheme,
    theme,
@@ -21,6 +24,7 @@ export const useThemeScheme = ({
 	}
 }
 
+/** @deprecated (no alternative) */
 export const useTitleThemeScheme = ({
 	scheme,
 	theme,
