+++ packages/ui/src/deprecated/Layout/Types.ts
@@ -1,5 +1,6 @@
 import { Intent, Scheme } from '../../types'
 
+/** @deprecated (no alternative) */
 export interface ThemeScheme {
 	scheme?: Scheme
 	theme?: Intent
