+++ packages/ui/src/deprecated/index.css
@@ -1,2 +1,3 @@
 /* DEPRECATED */
+@import "Layout";
 @import "Logo";
